@@ -3,28 +3,28 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve:admin": "cross-env VUE_CLI_SERVICE_CONFIG_PATH=$PWD/vue.entry.admin.config.js vue-cli-service serve",
-    "serve:mobile": "cross-env VUE_CLI_SERVICE_CONFIG_PATH=$PWD/vue.entry.mobile.config.js vue-cli-service serve",
     "build:admin": "cross-env VUE_CLI_SERVICE_CONFIG_PATH=$PWD/vue.entry.admin.config.js vue-cli-service build",
     "build:mobile": "cross-env VUE_CLI_SERVICE_CONFIG_PATH=$PWD/vue.entry.mobile.config.js vue-cli-service build",
     "inspect:admin": "cross-env VUE_CLI_SERVICE_CONFIG_PATH=$PWD/vue.entry.admin.config.js vue-cli-service inspect --mode production > webpack.entry.admin.js",
-    "inspect:mobile": "cross-env VUE_CLI_SERVICE_CONFIG_PATH=$PWD/vue.entry.mobile.config.js vue-cli-service inspect --mode production > webpack.entry.mobile.js"
+    "inspect:mobile": "cross-env VUE_CLI_SERVICE_CONFIG_PATH=$PWD/vue.entry.mobile.config.js vue-cli-service inspect --mode production > webpack.entry.mobile.js",
+    "serve:admin": "cross-env VUE_CLI_SERVICE_CONFIG_PATH=$PWD/vue.entry.admin.config.js vue-cli-service serve",
+    "serve:mobile": "cross-env VUE_CLI_SERVICE_CONFIG_PATH=$PWD/vue.entry.mobile.config.js vue-cli-service serve"
   },
   "dependencies": {
     "axios": "^0.21.0",
     "axios-extensions": "^3.1.3",
     "axios-progress-bar": "^1.2.0",
     "core-js": "^3.6.5",
-    "element-ui": "^2.14.0",
+    "element-plus": "^1.0.1-alpha.12",
     "perfect-scrollbar": "^1.5.0",
     "qs": "^6.9.4",
     "register-service-worker": "^1.7.1",
-    "vue": "^2.6.12",
+    "vue": "^3.0.2",
     "vue-context": "^6.0.0",
     "vue-cropperjs": "^4.1.0",
-    "vue-router": "^3.4.8",
+    "vue-router": "^v4.0.0-rc.1",
     "vue-router-back-button": "^1.3.0",
-    "vuex": "^3.5.1"
+    "vuex": "^v4.0.0-rc.1"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^4.5.8",
@@ -33,6 +33,7 @@
     "@vue/cli-plugin-router": "^4.5.8",
     "@vue/cli-plugin-vuex": "^4.5.8",
     "@vue/cli-service": "^4.5.8",
+    "@vue/compiler-sfc": "^3.0.2",
     "@vue/eslint-config-prettier": "^6.0.0",
     "assign-deep": "^1.0.1",
     "babel-eslint": "^10.1.0",
@@ -43,10 +44,7 @@
     "less": "^3.12.2",
     "less-loader": "^7.0.2",
     "mocker-api": "^2.7.0",
-    "prettier": "^2.1.2",
-    "sass": "^1.28.0",
-    "sass-loader": "^10.0.4",
-    "vue-template-compiler": "^2.6.12"
+    "prettier": "^2.1.2"
   },
   "eslintConfig": {
     "root": true,
@@ -55,9 +53,9 @@
     },
     "extends": [
       "plugin:vue/essential",
+      "eslint:recommended",
       "@vue/prettier"
     ],
-    "rules": {},
     "parserOptions": {
       "parser": "babel-eslint"
     }
@@ -65,7 +63,8 @@
   "prettier": {
     "tabWidth": 2,
     "semi": false,
-    "singleQuote": true
+    "singleQuote": true,
+    "quoteProps": "as-needed"
   },
   "postcss": {
     "plugins": {
@@ -74,6 +73,7 @@
   },
   "browserslist": [
     "> 1%",
-    "last 2 versions"
+    "last 2 versions",
+    "not dead"
   ]
 }
