@@ -1,9 +1,9 @@
 const PermissionCheck = {
-  install(Vue, { store }) {
-    Vue.prototype.$can = (actionId) => {
+  install(app, { store }) {
+    app.config.globalProperties.$can = (actionId) => {
       return store.getters['auth/isRouteInAcl'](actionId)
     }
   },
 }
 
-module.exports = PermissionCheck
+export default PermissionCheck
