@@ -19,7 +19,7 @@ export default {
       }
     }, 900)
   },
-  destroyed() {
+  unmounted() {
     if (this.spinnerTimeoutId) {
       clearTimeout(this.spinnerTimeoutId)
     }
@@ -27,12 +27,12 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="less">
 body {
   color: #647279;
   font-family: Helvetica Neue, Helvetica, PingFang SC, Hiragino Sans GB,
     Microsoft YaHei, SimSun, sans-serif;
 }
 
-@import '../../core/assets/styles/base.scss';
+@import '../../core/assets/styles/base.less';
 </style>
