@@ -82,7 +82,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="less">
 .placeholder-form {
   display: block;
 }
