@@ -23,18 +23,20 @@
         <header-tab></header-tab>
       </header>
       <main class="ag-main">
-        <transition name="fade">
-          <keep-alive :include="cachedTabs">
-            <router-view
-              :key="
-                $route.meta && $route.meta.parent
-                  ? $route.meta.parent.path
-                  : $route.path
-              "
-              v-if="isRouterAlive"
-            />
-          </keep-alive>
-        </transition>
+        <router-view
+          v-slot="{ Component }"
+          :key="
+            $route.meta && $route.meta.parent
+              ? $route.meta.parent.path
+              : $route.path
+          "
+          v-if="isRouterAlive">
+          <transition name="fade">
+            <keep-alive :include="cachedTabs">
+              <component :is="Component" />
+            </keep-alive>
+          </transition
+        ></router-view>
       </main>
     </section>
   </section>
@@ -89,7 +91,7 @@ export default {
 
       this.isRouterAlive = false
 
-      this.$nextTick(() => {
+      await this.$nextTick(() => {
         this.isRouterAlive = true
         this.$store.dispatch('tabs/updateCache', cacheNames)
       })
@@ -97,11 +99,12 @@ export default {
   },
 }
 </script>
-<style lang="scss">
-@import '../../../core/assets/styles/mixin';
-$asideWidth: 220px;
-$asideCollapsedWidth: 60px;
-$asideBackgroundColor: #2c343f;
+
+<style lang="less">
+@import '../../../core/assets/styles/mixin.less';
+@aside-width: 220px;
+@aside-collapsed-width: 60px;
+@aside-background-color: #2c343f;
 
 .fade-enter-active,
 .fade-leave-active {
@@ -121,15 +124,15 @@ $asideBackgroundColor: #2c343f;
 
   &.ag-collapsed {
     .ag-aside {
-      width: $asideCollapsedWidth;
+      width: @aside-collapsed-width;
     }
 
     .ag-layout {
-      margin-left: $asideCollapsedWidth;
+      margin-left: @aside-collapsed-width;
     }
 
     .ag-header {
-      width: calc(100% - #{$asideCollapsedWidth});
+      width: calc(~'100% - ' @aside-collapsed-width);
     }
   }
 
@@ -138,10 +141,10 @@ $asideBackgroundColor: #2c343f;
     box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.01);
     top: 0;
     left: 0;
-    width: $asideWidth;
+    width: @aside-width;
     z-index: 10;
     height: 100vh;
-    background-color: $asideBackgroundColor;
+    background-color: @aside-background-color;
 
     .ag-logo {
       text-align: center;
@@ -164,7 +167,7 @@ $asideBackgroundColor: #2c343f;
     }
 
     .el-menu {
-      background-color: $asideBackgroundColor;
+      background-color: @aside-background-color;
 
       .el-menu-item,
       .el-submenu__title {
@@ -179,7 +182,7 @@ $asideBackgroundColor: #2c343f;
 
   .ag-layout {
     position: relative;
-    margin-left: $asideWidth;
+    margin-left: @aside-width;
     display: flex;
     flex: auto;
     flex-direction: column;
@@ -188,14 +191,14 @@ $asideBackgroundColor: #2c343f;
 
   .ag-header {
     position: fixed;
-    width: calc(100% - #{$asideWidth});
+    width: calc(~'100% - ' @aside-width);
     z-index: 10;
     right: 0;
     overflow-x: hidden;
 
     .ag-nav {
+      .clearfix();
       background: #fff;
-      @include clearfix();
       box-shadow: 2px 2px 2px rgba(0, 0, 0, 0.02);
     }
   }
