@@ -2,7 +2,7 @@
   <div v-if="pages" class="ag-pages">
     <el-pagination
       :background="true"
-      :current-page.sync="pages.currentPage"
+      v-model:current-page="pages.currentPage"
       :page-sizes="[20, 50, 100, 200]"
       :page-size="pages.perPage"
       layout="total, prev, pager, next, sizes, jumper"
@@ -41,7 +41,7 @@ export default {
   },
 }
 </script>
-<style lang="scss">
+<style lang="less">
 .ag-pages {
   margin-top: 20px;
   text-align: right;
