@@ -1,6 +1,4 @@
-import Vue from 'vue'
-import VueRouter from 'vue-router'
-import VueRouterBackButton from 'vue-router-back-button'
+import { createRouter, createWebHistory } from 'vue-router'
 
 import notFound from '@admin/views/site/NotFound'
 import siteIndex from '@admin/views/site/Index'
@@ -18,8 +16,6 @@ import shopProductRoute from './shop/product'
 import districtRoute from './district/district'
 
 import AdminLayout from '@admin/components/AdminLayout'
-
-Vue.use(VueRouter)
 
 const routes = [
   {
@@ -65,7 +61,7 @@ const routes = [
       ...shopCategoryRoute,
       ...districtRoute,
       {
-        path: '*',
+        path: '/:pathMatch(.*)*',
         component: notFound,
         meta: {
           title: '页面未找到',
@@ -75,9 +71,8 @@ const routes = [
   },
 ]
 
-const router = new VueRouter({
-  base: '/admin',
-  mode: 'history',
+const router = createRouter({
+  history: createWebHistory('/admin'),
   routes: routes,
   scrollBehavior: (to, from, savedPosition) => {
     if (savedPosition) {
@@ -85,13 +80,11 @@ const router = new VueRouter({
     }
 
     if (to.hash) {
-      return { selector: to.hash }
+      return { el: to.hash, behavior: 'smooth' }
     }
 
-    return { x: 0, y: 0 }
+    return { left: 0, top: 0 }
   },
 })
-
-Vue.use(VueRouterBackButton, { router, ignoreRoutesWithSameName: false })
 
 export default router
