@@ -1,24 +1,24 @@
-import Vue from 'vue'
+import { useHttpClient } from '@core/plugins/HttpClient'
 
 const UserAddressService = {
   all(query) {
-    return Vue.http.get('user-addresses', query)
+    return useHttpClient().get('user-addresses', query)
   },
 
   delete(id) {
-    return Vue.http.delete('user-addresses/' + id)
+    return useHttpClient().delete('user-addresses/' + id)
   },
 
   create(userAddress) {
-    return Vue.http.post('user-addresses', userAddress)
+    return useHttpClient().post('user-addresses', userAddress)
   },
 
   view(id) {
-    return Vue.http.get('user-addresses/' + id)
+    return useHttpClient().get('user-addresses/' + id)
   },
 
   edit(userAddress) {
-    return Vue.http.put('user-addresses/' + userAddress.id, userAddress)
+    return useHttpClient().put('user-addresses/' + userAddress.id, userAddress)
   },
 }
 
