@@ -1,9 +1,6 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
+import { createStore } from 'vuex'
 import app from './app'
 
-Vue.use(Vuex)
-
-const store = new Vuex.Store(app)
+const store = createStore(app)
 
 export default store
