@@ -313,7 +313,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="less">
 .check-group-box {
   line-height: 1;
 
