@@ -26,7 +26,7 @@
       />
       <el-table-column prop="name" width="220" label="品牌名称" />
       <el-table-column width="110" align="center" label="图片">
-        <template slot-scope="scope">
+        <template v-slot="scope">
           <img
             :alt="scope.row.name"
             v-if="scope.row.logo"
@@ -42,7 +42,7 @@
         </template>
       </el-table-column>
       <el-table-column width="260" label="品牌分类">
-        <template slot-scope="scope">
+        <template v-slot="scope">
           <div class="brand-category-tags">
             <template v-for="shopCategory in scope.row.shopCategories">
               <el-tag size="medium" :key="scope.row.id + '-' + shopCategory.id">
@@ -53,7 +53,7 @@
         </template>
       </el-table-column>
       <el-table-column label="品牌介绍">
-        <template slot-scope="scope">
+        <template v-slot="scope">
           <div class="description">{{ scope.row.description }}</div>
           <div class="website" v-if="scope.row.website">
             网址：<a :href="scope.row.website" target="_blank">{{
@@ -63,7 +63,7 @@
         </template>
       </el-table-column>
       <el-table-column width="135" label="操作" fixed="right" align="right">
-        <template slot-scope="scope">
+        <template v-slot="scope">
           <el-button
             :disabled="!$can('shop-brand/edit')"
             type="primary"
@@ -164,7 +164,7 @@ export default {
   },
 }
 </script>
-<style lang="scss">
+<style lang="less">
 .shop-brands {
   .logo {
     border: 1px solid #dddddd;
